+++ FakeIt.Common/APIModel/ReadAPI/ReadAPIResponse.cs
@@ -1,5 +1,4 @@
 ﻿using FakeIt.Common.Common;
-using Newtonsoft.Json.Linq;
 
 namespace FakeIt.Common.APIModel.ReadAPI
 {
