+++ FakeIt.Common/Common/CommonConstants.cs
@@ -1,0 +1,7 @@
+﻿namespace FakeIt.Common.Common
+{
+    public static class CommonConstants
+    {
+        public static string INTERNAL_SERVER_ERROR = "Internal server error.";
+    }
+}
