+++ FakeIt.Web/Filters/CustomValidationFilter.cs
@@ -10,11 +10,11 @@ namespace FakeIt.Web.Filters
             if (!context.ModelState.IsValid)
             {
                 var errors = context.ModelState
-                    .Where(e => e.Value.Errors.Count > 0)
+                    .Where(e => e.Value?.Errors.Count > 0)
                     .Select(e => new
                     {
                         Field = e.Key,
-                        Errors = e.Value.Errors.Select(x => x.ErrorMessage).ToArray()
+                        Errors = e.Value?.Errors.Select(x => x.ErrorMessage).ToArray()
                     });
 
                 var errorResponse = new
